separate searchbar into separate component
